refactor: revisit scale configs
