+++ packages/encodable/src/types/scale/Nice.ts
@@ -1,0 +1,1 @@
+export type NiceTime = 'second' | 'minute' | 'hour' | 'day' | 'week' | 'month' | 'year';
